+++ precompiles/utils/macro/Cargo.toml
@@ -19,3 +19,4 @@ proc-macro2 = "1.0"
 quote = "1.0"
 sha3 = "0.8"
 syn = { version = "1.0", features = [ "extra-traits", "fold", "full", "visit" ] }
+case = "1.0"
